Review changes

+++ src/Converter.java
@@ -1,10 +1,10 @@
-
 public class Converter {
-    public int milesToInches(int miles){
+    public double milesToInches(double miles) {
         int inchesInAMile = 63360;
         return miles * inchesInAMile;
     }
-    public double areaOfCircle(double radius){
+
+    public double areaOfCircle(double radius) {
         return (radius * radius) * Math.PI;
     }
 }
